room/requirements.txt
aenum==3.1.15
allure-pytest==2.13.5
allure-python-commons==2.13.5
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
assertpy==1.1
astroid==2.15.8
attrs==24.2.0
beautifulsoup4==4.12.3
black==23.1.0
blinker==1.8.2
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
ConfigArgParse==1.7
cryptography==43.0.1
curlify==2.2.1
decorator==5.1.1
delayed-assert==0.3.6
dill==0.3.8
Django==4.2.16
django-bootstrap-v5==1.0.11
django-cleanup==8.1.0
django-debug-toolbar==4.4.6
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
exceptiongroup==1.2.2
execnet==2.1.1
exrex==0.11.0
Faker==12.0.1
fast-map==0.1.0
fastapi==0.114.1
fastapi-utils==0.7.0
flake8==6.0.0
Flask==3.0.3
Flask-BasicAuth==0.2.0
Flask-Cors==5.0.0
gevent==21.12.0
geventhttpclient==2.3.1
greenlet==1.1.2
gunicorn==23.0.0
h11==0.14.0
idna==3.8
influxdb-client==1.46.0
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpath-ng==1.6.1
lazy-object-proxy==1.10.0
locust==2.12.1
lxml==5.3.0
Markdown==3.7
MarkupSafe==2.1.5
mccabe==0.7.0
mixer==7.2.2
msgpack==1.1.0
mypy-extensions==1.0.0
numpy==2.1.1
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
platformdirs==4.3.2
pluggy==1.5.0
ply==3.11
prometheus_client==0.20.0
psutil==5.9.8
psycogreen==1.0.2
psycopg2-binary==2.9.9
PyByteBuffer==1.0.5
pycodestyle==2.10.0
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
pyflakes==3.0.1
PyJWT==2.9.0
pylint==2.17.0
PySocks==1.7.1
pytest==8.3.3
pytest-json-report==1.5.0
pytest-metadata==3.1.1
pytest-order==1.3.0
pytest-reportportal==5.1.5
pytest-rerunfailures==14.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyzmq==26.2.0
reactivex==4.0.4
reportportal-client==5.2.5
requests==2.32.3
roundrobin==0.0.4
selenium==4.21.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.1
starlette==0.38.5
tomli==2.0.1
tomlkit==0.13.2
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
unitest==1.4.34
urllib3==2.2.3
uvicorn==0.30.6
webdriver-manager==4.0.2
websocket-client==1.4.2
Werkzeug==3.0.4
wrapt==1.16.0
wsproto==1.2.0
zope.event==5.0
zope.interface==7.0.3
